Move discovery code to a module
