+++ tap_square/discover.py
@@ -1,0 +1,50 @@
+import datetime
+import json
+from singer import metadata
+import os
+from .streams import STREAMS
+
+def get_abs_path(path):
+    return os.path.join(os.path.dirname(os.path.realpath(__file__)), path)
+
+
+def get_schemas():
+
+    schemas = {}
+    schemas_metadata = {}
+
+    for stream_name, stream_metadata in STREAMS.items():
+
+        schema_path = get_abs_path('schemas/{}.json'.format(stream_name))
+        with open(schema_path) as file:
+            schema = json.load(file)
+
+        meta = metadata.get_standard_metadata(
+            schema=schema,
+            valid_replication_keys=stream_metadata.get('replication_keys', None),
+            replication_method=stream_metadata.get('replication_method', None)
+        )
+        schemas[stream_name] = schema
+        schemas_metadata[stream_name] = meta
+
+    return schemas, schemas_metadata
+
+
+def discover():
+
+    schemas, schemas_metadata = get_schemas()
+    streams = []
+
+    for schema_name, schema in schemas.items():
+        schema_meta = schemas_metadata[schema_name]
+
+        catalog_entry = {
+            'stream': schema_name,
+            'tap_stream_id': schema_name,
+            'schema': schema,
+            'metadata': schema_meta
+        }
+
+        streams.append(catalog_entry)
+
+    return {'streams': streams}
