+++ tap_square/streams.py
@@ -1,0 +1,6 @@
+STREAMS = {
+    'items': {
+        'replication_method': 'FULL_TABLE',
+        'replication_keys': ['updated_at']
+    }
+}
